+++ services/todo.service.ts
@@ -17,6 +17,12 @@ export const createTODOList = async ({ name, email }: { name: string; email: str
 };
 
 export const deleteTODOList = async (id: string) => {
+  const { error: todoTasksError } = await supabase.from('todo_tasks').delete().eq('todo_list_id', id);
+
+  if (todoTasksError) {
+    throw todoTasksError;
+  }
+
   const { error } = await supabase.from('todo_lists').delete().eq('id', id);
 
   if (error) {
